# OCR-Spring-Backend-main/swagger/ocr-api.yaml
openapi: 3.0.0
info:
  title: OCR API
  description: API for managing OCR apps
  version: 1.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://localhost
...
paths:
...
  /api/dashboard/percentages:
    get:
      tags: 
        - Dashboard 
      summary: Get the percentage of risk habits
      description: >
        This endpoint returns the percentage of risk habits. If no data is found, it returns an error message.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RiskHabitPercentage"
        "401":
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/dashboard/totals:
    get:
      tags: 
        - Dashboard
      summary: Get the total number of doctors, images and patients
      description: >
        This endpoint returns the total number of doctors, images and patients. If no data is found, it returns an error message.
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  doctors:
                    type: integer
                    description: Number of doctors
                    example: 10
                  images:
                    type: integer
                    description: Number of images
                    example: 100
                  patients:
                    type: integer
                    description: Number of patients
                    example: 1000
                    $ref: "#/components/schemas/TotalCounts"
        "401":
          description: Unauthorized access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/user/draftentry/add/{id}:
    post:
      tags: 
        - Draft Entry
      summary: Add a draft Entry
      description: >
        This endpoint adds a draft Entry. If the patient is not found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                start_time:
                  type: string
                  format: date-time
                end_time:
                  type: string
                  format: date-time
                complaint:
                  type: string
                findings:
                  type: string
                current_habits:
                  type: string
      responses:
        "200":
          description: Draft Entry added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftEntry"
        "401":
          description: Unauthorized access
        "404":
          description: Patient not registered
        "500":
          description: Internal Server Error
  /api/user/draftentry/get:
    get:
      tags: 
        - Draft Entry
      summary: Get all draft entries added by the user
      description: >
        This endpoint returns a list of all draft entries added by the user. It supports pagination and search.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: filter
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of draft entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DraftEntry"
        "401":
          description: Unauthorized access
        "500":
          description: Internal Server Error
  /api/user/draftentry/get/patient/{id}:
    get:
      tags: 
        - Draft Entry
      description: >
        This endpoint returns a list of all draft entries for a specific patient. It supports pagination and search.
      summary: Get all draft entries for a specific patient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: filter
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of draft entries for the patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DraftEntry"
        "401":
          description: Unauthorized access
        "500":
          description: Internal Server Error
  /api/user/draftentry/get/{id}:
    get:
      tags: 
        - Draft Entry
      summary: Get details of a specific draft entry
      description: >
        This endpoint returns the details of a specific draft entry. If the draft entry is not found, it returns an error message.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Details of the draft entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftEntry"
        "401":
          description: Unauthorized access
        "404":
          description: Draft entry not found
        "500":
          description: Internal Server Error
  /api/user/entry/add/{id}:
    post:
      tags:
        - Entry
      summary: Add a telecunsultation entry
      description: >
        This endpoint adds a Entry. If the patient is not found, it returns an error message.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The patient id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                  start_time:
                    type: string
                    format: date-time
                  end_time:
                    type: string
                    format: date-time
                  complaint:
                    type: string
                  findings:
                    type: string
                  current_habits:
                    type: string  
      responses:
        "200":
          description: telecunsultation entry created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeleConEntry" 
        "401":
          description: Unauthorized access
        "404" :
          description: Patient not found
        "500":
          description: Internal server error    
  /api/user/entry/get:
    get:
      tags:
        - Entry
      summary: get all entries added by user
      description: >
        This endpoint returns a list of all teleconsultation entries added by the user. It supports pagination and search.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: page number for pagination
        - in: query
          name: filter
          schema:
            type: string
            enum: [created Date,updated Date,Assigned,Unassigned,Reviewed,Unreviewed]
          description: filter for entries
      responses:
        "200":
          description: A list of telecunsultation entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeleConEntry"
        "401":
          description: Unauthorized access
        "500":  
          description: Internal server error
  /api/user/entry/get/patient/{id}:
    get:
      tags:
        - Entry
      summary: Get patient's entries
      description: >
        This endpoint returns a list of all teleconsultation entries for a specific patient. It supports pagination and search.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The patient Id
        - in: query
          name: page
          schema:
            type: integer
          description: page number for pagination
        - in: query
          name: filter
          schema:
            type: string
            enum: [created Date,updated Date]
          description: filter for entries
      responses:
        "200":
          description: A list of telecunsultation entries for the patient
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeleConEntry"
        "401":
          description: Unauthorized access
        "500":
          description: Internal server error
...
  /api/user/entry/get/{id}:
    get:
      tags:
        - Entry
      summary: get one entry details added by users
      description: >
        This endpoint returns the details of a Entry. If the entry is not found, it returns an error message.
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The entry Id
       
      responses:
        "200":
          description: telecunsultation entry details
          content:
            application/json:
              schema:
               $ref: "#/components/schemas/TeleConEntry"
        "401":
          description: Unauthorized access
        "500":
          description: Internal server error
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        message:
          type: string
          description: Error message
          example: "Internal Server Error!"
        error:
          type: string
          description: Detailed error information
          example: "Detailed error information here."
...
    RiskHabitPercentage:
      type: object
      properties:
        item:
          type: string
          description: Name of the risk habit
          example: Smoking
        percentage:
          type: number
          description: >
            Percentage of patients with this habit.
            Example: If there are 100 patients and 20 of them have the 'Smoking' habit, the percentage is 20%.
          example: 20
    TotalCounts:
      type: object
      properties:
        doctors:
          type: integer
          description: Total number of doctors
          example: 10
        patients:
          type: integer
          description: Total number of patients
          example: 50
        images:
          type: integer
          description: Total number of images
          example: 100    
...
    DraftEntry:
      type: object
      properties:
        patient:
          type: string
        clinician_id:
          type: string
        start_time:
          type: string
          format: date-time
        end_time:
          type: string
          format: date-time
        complaint:
          type: string
        findings:
          type: string
        current_habits:
          type: string
        images:
          type: array
          items:
            type: string
        reports:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TeleConEntry:
      type: object
      properties:
        patient:
          type: string
          description: ObjectId reference of the patient
        clinician_id:
          type: string
          description: ObjectId reference to the user
        complaint:
          type: string
          description: complaint of the patient
        start_time:
          type: string
          format: date-time
          description: start time of the telecon
        end_time:
          type: string
          format: date-time
          description: end time of the telecon
        findings:
          type: string
          description: findings of the telecon
        curent_habits:
          type: array
          items:
            type: string
          description: current habits of the patient
          default: []
        updated:
          type: boolean
          description: whether the telecon is updated
        reviewers:
          type: array
          items:
            type: string
            description: ObjectId reference to the User (reviewers)
        reviews:
          type: array
          items:
            type: string
            description: ObjectId reference to the Review
        images:
          type: array
          items:
            type: string
            description: ObjectId reference to the Image
        reports:
          type: array
          items: 
            type: string
            description: ObjectId reference to the Report
      required:
        - patient
        - clinician_id
...